unit-test equality up to optimization

+++ +OnlineCalibration/+aux/edgeSobelXY.m
@@ -9,6 +9,6 @@ mask = zeros(size(Ix));
 mask((1+marginPixels):end-marginPixels,(1+marginPixels):end-marginPixels) = 1;
 Ix(~mask) = 0;
 Iy(~mask) = 0;
-E = single(sqrt(Ix.^2+Iy.^2));
+E = double(sqrt(Ix.^2+Iy.^2));
 end
